backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
python-multipart==0.0.6
pillow==10.0.1
numpy==1.25.2
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.1.78
pydantic==2.4.2
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
asyncpg==0.28.0
python-dotenv==1.0.0 